+++ metadata.rb
@@ -5,7 +5,6 @@ license           'Apache-2.0'
 description       'Manages the ruby-build framework and its installed rubies. A LWRP is also defined.'
 source_url        'https://github.com/sous-chefs/ruby_build'
 issues_url        'https://github.com/sous-chefs/ruby_build/issues'
-long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 chef_version      '>= 13.0'
 version           '1.2.0'