+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This file is used to list changes made in each version of the ruby_build cookboo
 
 ## unreleased
 
-- get tests working
+- Fix CircleCI testing, bring it up to Sous-Chefs standards
+- Fix Markdown
+- Fix YAML
 
 ## 1.2.0 (2019-23-01)